Fix backward incompatibility * Change back the original protected methods * Change back the runtimeclaims' and runtimeparameters' parent classes * Change the JsHeaders class to original name
