+++ components/authentication-framework/org.wso2.carbon.identity.application.authentication.framework/src/main/java/org/wso2/carbon/identity/application/authentication/framework/config/model/graph/js/JsServletRequest.java
@@ -128,7 +128,7 @@ public abstract class JsServletRequest
 
     }
 
-    private HttpServletRequest getRequest() {
+    protected HttpServletRequest getRequest() {
 
         TransientObjectWrapper<HttpServletRequest> transientObjectWrapper = getWrapped();
         return transientObjectWrapper.getWrapped();
