+++ components/authentication-framework/org.wso2.carbon.identity.application.authentication.framework/src/main/java/org/wso2/carbon/identity/application/authentication/framework/config/model/graph/js/graaljs/JsGraalAuthenticatedUser.java
@@ -70,4 +70,8 @@ public class JsGraalAuthenticatedUser extends JsAuthenticatedUser implements Pro
         setMember(name, valueAsString);
     }
 
+    public boolean hasMember(String name) {
+
+        return true;
+    }
 }
