+++ components/authentication-framework/org.wso2.carbon.identity.application.authentication.framework/src/main/java/org/wso2/carbon/identity/application/authentication/framework/config/model/graph/js/graaljs/JsGraalAuthenticationContext.java
@@ -64,4 +64,9 @@ public class JsGraalAuthenticationContext extends JsAuthenticationContext implem
 
         return super.removeMemberObject(name);
     }
+
+    public boolean hasMember(String name) {
+
+        return true;
+    }
 }
