+++ components/authentication-framework/org.wso2.carbon.identity.application.authentication.framework/src/main/java/org/wso2/carbon/identity/application/authentication/framework/config/model/graph/js/graaljs/JsGraalClaims.java
@@ -53,4 +53,9 @@ public class JsGraalClaims extends JsClaims implements ProxyObject {
         String valueAsString = value.isString() ? value.asString() : String.valueOf(value);
         setMemberObject(claimUri, valueAsString);
     }
+
+    public boolean hasMember(String name) {
+
+        return true;
+    }
 }
