+++ components/authentication-framework/org.wso2.carbon.identity.application.authentication.framework/src/main/java/org/wso2/carbon/identity/application/authentication/framework/config/model/graph/js/graaljs/JsGraalCookie.java
@@ -51,4 +51,9 @@ public class JsGraalCookie extends JsCookie implements ProxyObject {
         String valueAsString = value.isString() ? value.asString() : String.valueOf(value);
         super.setMember(key, valueAsString);
     }
+
+    public boolean hasMember(String name) {
+
+        return true;
+    }
 }
