+++ components/authentication-framework/org.wso2.carbon.identity.application.authentication.framework/src/main/java/org/wso2/carbon/identity/application/authentication/framework/config/model/graph/js/graaljs/JsGraalParameters.java
@@ -58,4 +58,8 @@ public class JsGraalParameters extends JsParameters implements ProxyObject {
         return false;
     }
 
+    public boolean hasMember(String name) {
+
+        return true;
+    }
 }
