+++ components/authentication-framework/org.wso2.carbon.identity.application.authentication.framework/src/main/java/org/wso2/carbon/identity/application/authentication/framework/config/model/graph/js/graaljs/JsGraalServletRequest.java
@@ -68,4 +68,9 @@ public class JsGraalServletRequest extends JsServletRequest implements ProxyObje
 
         return false;
     }
+
+    public boolean hasMember(String name) {
+
+        return true;
+    }
 }
