+++ components/authentication-framework/org.wso2.carbon.identity.application.authentication.framework/src/main/java/org/wso2/carbon/identity/application/authentication/framework/config/model/graph/js/graaljs/JsGraalServletResponse.java
@@ -62,4 +62,9 @@ public class JsGraalServletResponse extends JsServletResponse implements ProxyOb
         String valueAsString = value.isString() ? value.asString() : String.valueOf(value);
         super.setMember(key, valueAsString);
     }
+
+    public boolean hasMember(String name) {
+
+        return true;
+    }
 }
