+++ components/authentication-framework/org.wso2.carbon.identity.application.authentication.framework/src/main/java/org/wso2/carbon/identity/application/authentication/framework/config/model/graph/js/graaljs/JsGraalStep.java
@@ -59,4 +59,9 @@ public class JsGraalStep extends JsStep implements ProxyObject {
         super.removeMemberObject(name);
         return true;
     }
+
+    public boolean hasMember(String name) {
+
+        return true;
+    }
 }
