+++ components/authentication-framework/org.wso2.carbon.identity.application.authentication.framework/src/main/java/org/wso2/carbon/identity/application/authentication/framework/config/model/graph/js/graaljs/JsGraalSteps.java
@@ -43,4 +43,9 @@ public class JsGraalSteps extends JsSteps implements ProxyArray {
     public void set(long index, Value value) {
         //Steps can not be set with script.
     }
+
+    public boolean hasMember(String name) {
+
+        return true;
+    }
 }
