+++ components/authentication-framework/org.wso2.carbon.identity.application.authentication.framework/src/main/java/org/wso2/carbon/identity/application/authentication/framework/config/model/graph/js/nashorn/JsNashornRuntimeClaims.java
@@ -18,7 +18,7 @@
 
 package org.wso2.carbon.identity.application.authentication.framework.config.model.graph.js.nashorn;
 
-import org.wso2.carbon.identity.application.authentication.framework.config.model.graph.js.JsRuntimeClaims;
+import org.wso2.carbon.identity.application.authentication.framework.config.model.graph.js.base.JsBaseRuntimeClaims;
 import org.wso2.carbon.identity.application.authentication.framework.context.AuthenticationContext;
 import org.wso2.carbon.identity.application.authentication.framework.model.AuthenticatedUser;
 
@@ -26,15 +26,38 @@ import org.wso2.carbon.identity.application.authentication.framework.model.Authe
  * Represent the user's runtime claims.
  * This wrapper uses jdk.nashorn engine.
  */
-public class JsNashornRuntimeClaims extends JsRuntimeClaims implements AbstractJsObject {
+public class JsNashornRuntimeClaims extends JsNashornClaims implements JsBaseRuntimeClaims, AbstractJsObject {
 
     public JsNashornRuntimeClaims(AuthenticationContext context, int step, String idp) {
 
-        super(context, step, idp);
+        super(context, step, idp, false);
     }
 
     public JsNashornRuntimeClaims(AuthenticationContext context, AuthenticatedUser user) {
 
-        super(context, user);
+        super(context, user, false);
+    }
+
+    public Object getMember(String claimUri) {
+
+        if (authenticatedUser != null) {
+            return getRuntimeClaim(claimUri);
+        }
+        return null;
+    }
+
+    public boolean hasMember(String claimUri) {
+
+        if (authenticatedUser != null) {
+            return hasRuntimeClaim(claimUri);
+        }
+        return false;
+    }
+
+    public void setMember(String claimUri, Object claimValue) {
+
+        if (authenticatedUser != null) {
+            setRuntimeClaim(claimUri, claimValue);
+        }
     }
 }
