+++ components/authentication-framework/org.wso2.carbon.identity.application.authentication.framework/src/main/java/org/wso2/carbon/identity/application/authentication/framework/config/model/graph/js/openjdk/nashorn/JsOpenJdkNashornRuntimeClaims.java
@@ -18,8 +18,7 @@
 
 package org.wso2.carbon.identity.application.authentication.framework.config.model.graph.js.openjdk.nashorn;
 
-import org.wso2.carbon.identity.application.authentication.framework.config.model.graph.js.JsRuntimeClaims;
-import org.wso2.carbon.identity.application.authentication.framework.config.model.graph.js.nashorn.AbstractJsObject;
+import org.wso2.carbon.identity.application.authentication.framework.config.model.graph.js.base.JsBaseRuntimeClaims;
 import org.wso2.carbon.identity.application.authentication.framework.context.AuthenticationContext;
 import org.wso2.carbon.identity.application.authentication.framework.model.AuthenticatedUser;
 
@@ -27,15 +26,39 @@ import org.wso2.carbon.identity.application.authentication.framework.model.Authe
  * Represent the user's runtime claims.
  * Since Nashorn is deprecated in JDK 11 and onwards. We are introducing OpenJDK Nashorn engine.
  */
-public class JsOpenJdkNashornRuntimeClaims extends JsRuntimeClaims implements AbstractJsObject {
+public class JsOpenJdkNashornRuntimeClaims extends JsOpenJdkNashornClaims
+        implements JsBaseRuntimeClaims, AbstractOpenJdkNashornJsObject {
 
     public JsOpenJdkNashornRuntimeClaims(AuthenticationContext context, int step, String idp) {
 
-        super(context, step, idp);
+        super(context, step, idp, false);
     }
 
     public JsOpenJdkNashornRuntimeClaims(AuthenticationContext context, AuthenticatedUser user) {
 
-        super(context, user);
+        super(context, user, false);
+    }
+
+    public Object getMember(String claimUri) {
+
+        if (authenticatedUser != null) {
+            return getRuntimeClaim(claimUri);
+        }
+        return null;
+    }
+
+    public boolean hasMember(String claimUri) {
+
+        if (authenticatedUser != null) {
+            return hasRuntimeClaim(claimUri);
+        }
+        return false;
+    }
+
+    public void setMember(String claimUri, Object claimValue) {
+
+        if (authenticatedUser != null) {
+            setRuntimeClaim(claimUri, claimValue);
+        }
     }
 }
